+++ Phonebook-master/Phonebook.Tests/GetAllTest.cs
@@ -1,0 +1,12 @@
+using NUnit.Framework;
+
+namespace PhonebookTest;
+
+public class GetAllTest
+{
+    [Test]
+    public void GetAll_AllSubscribers_SuccessfullyReceived()
+    {
+        
+    }
+}
